move conflict nets processing to PartialRouter
